requirements.txt
openai
python-dotenv
PyYAML
PyMuPDF
tqdm
